David control

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: acombier <acombier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/05 12:11:07 by acombier          #+#    #+#             */
-/*   Updated: 2025/11/14 18:23:01 by acombier         ###   ########.fr       */
+/*   Updated: 2025/11/19 14:17:28 by acombier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
